borrowing copy possible, updates number of copies in frontend, solved unit tests fails in copycommandresttest

+++ src/main/java/nl/sogyo/library/services/rest/libraryapi/json/message/AddCopyMessage.java
@@ -2,22 +2,22 @@ package nl.sogyo.library.services.rest.libraryapi.json.message;
 
 public class AddCopyMessage extends Message {
 	
-	private int copiesOfBook;
+	private long copiesOfBook;
 	private static final String messagePositive = "Exemplaar is toegevoegd aan de database";
 	private static final String messageNegative = "Exemplaar toevoegen mislukt";
 	
 	public AddCopyMessage() {}
 	
-	public AddCopyMessage(boolean commandSucceeded, int copiesOfBook) {
+	public AddCopyMessage(boolean commandSucceeded, long copiesOfBook) {
 		super(commandSucceeded, messagePositive, messageNegative);
 		this.copiesOfBook = copiesOfBook;
 	}
 	
-	public int getCopiesOfBook() {
+	public long getCopiesOfBook() {
 		return copiesOfBook;
 	}
 	
-	public void setCopiesOfBook(int copiesOfBook) {
+	public void setCopiesOfBook(long copiesOfBook) {
 		this.copiesOfBook = copiesOfBook;
 	}
 }
