+++ src/main/java/nl/sogyo/library/services/rest/libraryapi/json/message/DeleteCopyMessage.java
@@ -2,22 +2,22 @@ package nl.sogyo.library.services.rest.libraryapi.json.message;
 
 public class DeleteCopyMessage extends Message {
 	
-	private int copiesOfBook;
+	private long copiesOfBook;
 	private static final String messagePositive = "Exemplaar is verwijderd uit de database";
 	private static final String messageNegative = "Exemplaar verwijderen mislukt";
 	
 	public DeleteCopyMessage() {}
 	
-	public DeleteCopyMessage(boolean commandSucceeded, int copiesOfBook) {
+	public DeleteCopyMessage(boolean commandSucceeded, long copiesOfBook) {
 		super(commandSucceeded, messagePositive, messageNegative);
 		this.copiesOfBook = copiesOfBook;
 	}
 
-	public int getCopiesOfBook() {
+	public long getCopiesOfBook() {
 		return copiesOfBook;
 	}
 	
-	public void setCopiesOfBook(int copiesOfBook) {
+	public void setCopiesOfBook(long copiesOfBook) {
 		this.copiesOfBook = copiesOfBook;
 	}
 }
